Add deep hashing for protocols
This is needed to assure that whole protocols
can be compared not equal by just comparing their
hashes. This implies that the hash needs to depend
on the hashes of all members of the protocol and
the procol id. Protocol hashes are calculated at
runtime in Python and at compile time in C++.
Computing protocol hashes at generation time was
rejected due to the need to keep protocol and
types decoupled (protocol and types are generated
in possibly two separate runs of the messgen).